requirements.txt
arch==7.2.0
beautifulsoup4==4.13.4
cloudpickle==2.2.1
hanlp==2.1.1
mlflow==2.21.3
mlflow_skinny==2.21.3
numpy==1.24.4
pandas==1.5.3
pandas_ta==0.3.14b0
PyMySQL==1.1.1
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
tensorflow==2.19.0
torch==2.6.0+cu126
transformers==4.50.3
yfinance==0.2.55
